refactor(discount): discount 상수 분류 및 패키지 구조 리팩토링
