+++ src/main/java/christmas/domain/discount/DiscountManager.java
@@ -1,0 +1,4 @@
+package christmas.domain.discount;
+
+public final class DiscountManager {
+}
