+++ src/main/java/christmas/domain/discount/strategies/Gift.java
@@ -1,13 +1,13 @@
-package christmas.domain.discount;
+package christmas.domain.discount.strategies;
 
-import christmas.domain.OrderProduct;
+import static christmas.domain.discount.constants.DiscountInfo.DEFAULT_GIFT_QUANTITY;
+import static christmas.domain.discount.constants.DiscountInfo.GIFT_DISCOUNT_STRING;
+
 import christmas.domain.Orders;
+import christmas.domain.discount.Discount;
 import christmas.domain.vo.Product;
-import org.junit.jupiter.api.Order;
 
 public class Gift implements Discount {
-    private static final String DEFAULT_GIFT_QUANTITY = "1개";
-    private static final String GIFT_DISCOUNT_STRING = "증정 이벤트";
     private final boolean isGiftAvailable;
 
     public Gift(boolean isGiftAvailable) {
