+++ src/main/java/christmas/domain/discount/strategies/WeekendDiscount.java
@@ -1,29 +1,29 @@
-package christmas.domain.discount;
+package christmas.domain.discount.strategies;
 
 import static christmas.constants.Number.EVENT_MONTH;
 import static christmas.constants.Number.EVENT_YEAR;
+import static christmas.domain.discount.constants.DiscountInfo.WEEKEND_DISCOUNT_PRICE;
+import static christmas.domain.discount.constants.DiscountInfo.WEEKEND_DISCOUNT_STRING;
 
 import christmas.domain.Orders;
+import christmas.domain.discount.Discount;
 import christmas.domain.vo.MenuType;
 import java.time.DayOfWeek;
 import java.time.LocalDate;
 
-public class WeekdayDiscount implements Discount {
-    private static final int WEEKDAY_DISCOUNT_PRICE = 2023;
-    private static final String WEEKDAY_DISCOUNT_STRING = "평일 할인";
-
+public class WeekendDiscount implements Discount {
     @Override
     public int applyDiscount(Orders orders) {
         int orderDay = orders.getDate();
         if (isWeekend(orderDay)) {
-            return 0;
+            return getDiscountPrice(orders);
         }
-        return getDiscountPrice(orders);
+        return 0;
     }
 
     @Override
     public String getDiscountName() {
-        return WEEKDAY_DISCOUNT_STRING;
+        return WEEKEND_DISCOUNT_STRING;
     }
 
     public boolean isWeekend(int orderDay) {
@@ -34,7 +34,7 @@ public class WeekdayDiscount implements Discount {
     }
 
     private int getDiscountPrice(Orders orders) {
-        int desertMenuCount = orders.findCountMenuType(MenuType.DESSERT);
-        return desertMenuCount * WEEKDAY_DISCOUNT_PRICE;
+        int mainMenuCount = orders.findCountMenuType(MenuType.MAIN);
+        return mainMenuCount * WEEKEND_DISCOUNT_PRICE;
     }
 }
