+++ src/main/java/christmas/dto/BenefitDto.java
@@ -1,6 +1,6 @@
 package christmas.dto;
 
-import christmas.domain.discount.Gift;
+import christmas.domain.discount.strategies.Gift;
 
 public record BenefitDto(int preTotalPrice, Gift gift) {
     public String getGiftMessage() {
