+++ src/main/java/christmas/service/ConvertService.java
@@ -1,7 +1,4 @@
 package christmas.service;
 
-import christmas.domain.discount.Gift;
-import christmas.dto.BenefitDto;
-
 public class ConvertService {
 }
